[#29689] Add 'exectool' in configuration
Former-commit-id: 7e5b7fc296427c596230a1f23819e66d265ae5f8

+++ run_aster/run_aster_main.py
@@ -234,7 +234,7 @@ def main(argv=None):
     opts["interactive"] = args.interactive
     calc = RunAster.factory(export, **opts)
     status = calc.execute(args.wrkdir)
-    if tmpf:
+    if tmpf and not opts["env"]:
         os.remove(tmpf)
     return status.exitcode
